requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv
google-generativeai==0.3.2
httpx==0.25.0